+++ esta-webjs-generator/service/index.js
@@ -5,52 +5,33 @@
 'use strict';
 
 var generators = require('yeoman-generator');
+var mkdirp = require('mkdirp');
+var util = require('../util');
 
-function upperCase(string) {
-    return string.charAt(0).toUpperCase() + string.slice(1);
-}
-
-function lowerCase(string) {
-    return string.charAt(0).toLowerCase() + string.slice(1);
-}
-
-var self;
+var yo;
 
 module.exports = generators.Base.extend({
     constructor: function () {
 
-        self = this;
+        yo = this;
 
-        generators.Base.apply(self, arguments);
+        generators.Base.apply(yo, arguments);
 
-        self.argument('name', {
+        yo.argument('name', {
             type: String,
             required: true,
-            description: 'Name of service'
+            description: 'Name of Service'
         });
     },
 
     writing: function () {
-        var nameUpperCase = upperCase(self.name);
-        var nameLowerCase = lowerCase(self.name);
-
         var files = [
             'replace___me.service.js',
             'replace___me.js',
             'replace___me.spec.js'
         ];
 
-        files.forEach(function(file){
-            var newFilename = file.replace(/replace___me/, nameLowerCase);
-
-            self.fs.copy(self.templatePath(file), self.destinationPath(newFilename), {
-                process: function (content) {
-                    return content.toString()
-                        .replace(/Replace___me/g, nameUpperCase)
-                        .replace(/replace___me/g, nameLowerCase);
-                }
-            });
-        });
+        util.create(yo, 'service', files);
     }
 
 });
